Rename GRadioGroupBtn to GSegmentedControl
