+++ src/PACKAGE/components/GSegmentedControl.svelte
